requirements.txt
dj-database-url==0.3.0
Django==1.9.1
django-filter==0.12.0
djangorestframework==3.3.2
gunicorn==19.4.5
Markdown==2.6.5
psycopg2==2.6.1
wheel==0.24.0
whitenoise==2.0.6
